Create SVC_reflectance_treatment
Created Leaf_level folder for all measurements taken by hand-held instruments. The first file is by Chris Wong: "SVC_reflectance_treatment_week4.R". This is the original script with no additions/annotations from Jessie. There is a lot of code for plotting the data at the end of the script that may not be needed for the wrangling.