Create SVC_reflectance_treatment.R
Created new folder for leaf level measurements. This folder will contain all files associated with hand-held instruments such as SVC and the LI-600. The first file included is by Chris Wong: "SVC_reflectance_treatment_week4.R". This is the original script sent to Jessie and there are no edits or additions. The second half of the script is code for plotting the data. This may get deleted.