Reshaping meta data file for merging with data
The meta data file contains informaton about the plantS. For example, you may have a number of plants, under different treatments (e.g. different pathogens) and there are different genotypes. As this is the file that is typically used for recordind data in the field, it might not be the best layout for merging data together. Here we define the differentiating columns, and then reshape the data so that there is a row for every measurement, not just per plant.

The number of leaves measured per plant can differ by project. I created a for loop so that R will automatically for all columns with the prefix "SVC_" and then reshape the data accordingly.