Merge dataframes and summarise spectral data
There are two files, one with all the data entered (meta data) and another with all of the spectral data. We need to merge these files together so that we know which spectrum belongs to which plant.

Using dplyr, we left join the spectral data onto the meta data. We then summarised the data to get an mean and standard deviation of reflectance per wavelenght, per plant.