Create script for LI-600
We use the Li-600 to get plant physiology information. We use it to collect measurements on stomatal conductance and fluorescence. There two types of fluorescence measurements, one collected during the daytime(light-adapted leaf) and one collected during the night (dark adapted leaf). This script will be used for importing the raw data, merging it with meta data files and exporting it for analyses.