importing Li-600 data and wrangling meta data
Added code for importing Li-600 data. Will need to edit this so that people don't need to assign a file path. The tricky part is the file path isn't constant because the exported folder name changes with the date of the measurement.

Added code to wrangle meta data. Same code as that in SVC_wrangle