Adding all wrangling code from previous script.
This script is the base file for preparing the Li-600 data. It currently is a script as there are inconsistencies with how the data was collected initially. After a meeting discussing the configuration settings of the Li-600 for dark-adapted measurements, the filing syster/ names of variables of interest should be consistent and we can turn this script in a series of functions for data preparation