Code for import data, wrangling date data
Reading in data is fastest using the data.table package and using function fread.
Data is imported with values in one object and time information in another.

Time is found in column V3. Year and day-of-year is pasted together and so we need to separate this column to summarise the data. We use the lubridate package to create a "date" column and then spplit up the month and year for summarising later