creating a manipulated dataframe and summarised.
I have manipulated the raw dataframe so that it is cleaned up and scaled for summarising. I currently have the summarised df as separate code to the manipulation. I think I will join these together so that we don't have have multiple copies of the same dataframe which takes up a lot of space.