Added code to extract SVC data from multiple folders.
This code is using tidyverse to extract SVC data stored in multiple columns. We first get all the folder directories, then we select only those folders with svc data using a unique identifier. Then we use Purrr to read in all the .sig files and append them together in a single dataframe.

The final step which I'm not sure about is how to perform the interpolation of the spectral data when we convert the wavelengths from decimals to integers. This is likely to require our own function to be written as the as.rspec doesn't seem to work with tidyverse setup